requirements.txt
google-cloud-storage
google-cloud-vision
python-dotenv
